pyproject.toml
[project]
name = "rcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.10",
]

[dependency-groups]
dev = [
    "faker>=37.8.0",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "pytest-faker>=2.0.0",
]
